chore(deps): upgrade biome 2, vitest 4, ts 6, chai 6, @noble/hashes 2
Major-version bumps across all dev tooling and one runtime dep:

  @noble/hashes  1.8 → 2.2   import paths consolidated (legacy.js + sha2.js, .js suffix required)
  @biomejs/biome 1.9 → 2.4   config migrated to v2 schema (includes/assist/noConsole)
  vitest         2.1 → 4.1   workspace mechanism removed → projects folded into config files
  typescript     5.9 → 6.0   ignoreDeprecations=6.0 added for tsup's internal baseUrl
  chai           4.5 → 6.2   no API changes; assert surface used in tests unchanged
  tsup           8.3 → 8.5
  @types/node    20.x → 22.x LTS line matching engines >=20

Notable touchpoints:

- src/crypto/backend.web.ts: sha1/ripemd160/md5 now from @noble/hashes/legacy.js,
  sha2.* from @noble/hashes/sha2.js
- biome.json: rewritten to v2 shape; noConsole rule configured with allow=[warn,error]
  to preserve v1 noConsoleLog intent (the auto-migration inverted the semantics)
- vitest.workspace.ts + vitest.bench.workspace.ts deleted; replaced with
  vitest.config.ts + vitest.bench.config.ts using test.projects
- package.json bench script: --workspace= → --config=
- tsconfig.json: ignoreDeprecations "6.0" silences tsup DTS baseUrl warning
- tsconfig.test.json: include list updated to new config files
- src/node-rsa.ts: catch (e) → catch (×2) for biome 2's stricter unused-var rule
- test/node-rsa.spec.ts: drop unused fromBase64 import

Format/import-ordering churn in bench/, src/{formats,index.*,rsa,schemes}, and
test/{formats,schemes} comes from biome 2's organizeImports being stricter than v1.

Bundle-size budgets in scripts/check-bundle-size.mjs raised to accommodate
existing pre-bump bundle (the raw size was already over the prior 105 KB
browser budget; gzip budget unchanged at 30 KB).

Verified: typecheck, lint, 1613 tests across node + browser-emulated, build
(ESM + CJS + DTS), and bench (node + js-jsbn + js-native) all green.
