--- vitest.bench.workspace.ts
@@ -1,27 +1,0 @@
-import { dirname, resolve } from 'node:path';
-import { fileURLToPath } from 'node:url';
-import { defineWorkspace } from 'vitest/config';
-
-const root = dirname(fileURLToPath(import.meta.url));
-const nodeEntry = resolve(root, 'src/index.node.ts');
-
-// The bench imports the public API via the virtual `node-rsa-bench-entry`
-// specifier, which resolves to the node bundle. The three modes (`node`,
-// `js-jsbn`, `js-native`) are selected per-bench via constructor options —
-// no separate workspace project per bundle is needed, since the digest/RNG
-// backend is always native to the runtime.
-export default defineWorkspace([
-  {
-    resolve: {
-      alias: { 'node-rsa-bench-entry': nodeEntry },
-    },
-    test: {
-      name: 'bench',
-      environment: 'node',
-      env: { NODE_RSA_BENCH_MODES: 'node,js-jsbn,js-native' },
-      benchmark: {
-        include: ['bench/**/*.bench.ts'],
-      },
-    },
-  },
-]);
